+++ packages/block-editor/src/components/block-mover/index.js
@@ -19,12 +19,7 @@ import BlockDraggable from '../block-draggable';
 import { BlockMoverUpButton, BlockMoverDownButton } from './button';
 import { store as blockEditorStore } from '../../store';
 
-function BlockMover( {
-	clientIds,
-	hideDragHandle,
-	isBlockMoverUpButtonDisabled,
-	isBlockMoverDownButtonDisabled,
-} ) {
+function BlockMover( { clientIds, hideDragHandle } ) {
 	const {
 		canMove,
 		rootClientId,
@@ -104,7 +99,6 @@ function BlockMover( {
 					<ToolbarItem>
 						{ ( itemProps ) => (
 							<BlockMoverUpButton
-								disabled={ isBlockMoverUpButtonDisabled }
 								clientIds={ clientIds }
 								{ ...itemProps }
 							/>
@@ -113,7 +107,6 @@ function BlockMover( {
 					<ToolbarItem>
 						{ ( itemProps ) => (
 							<BlockMoverDownButton
-								disabled={ isBlockMoverDownButtonDisabled }
 								clientIds={ clientIds }
 								{ ...itemProps }
 							/>