+++ packages/components/src/toggle-group-control/test/index.tsx
@@ -353,6 +353,60 @@ describe.each( [
 		);
 	}
 
+	it( 'should render the label', () => {
+		render(
+			<Component label="Test Toggle Group Control">{ options }</Component>
+		);
+
+		expect( screen.getByText( 'Test Toggle Group Control' ) ).toBeVisible();
+	} );
+
+	it( 'should still label the control accessibly when hideLabelFromVision is true', () => {
+		render(
+			<Component label="Test Toggle Group Control" hideLabelFromVision>
+				{ options }
+			</Component>
+		);
+
+		expect(
+			screen.getByRole( 'radiogroup', {
+				name: 'Test Toggle Group Control',
+			} )
+		).toBeVisible();
+	} );
+
+	it( 'should accessibly associate the help text', () => {
+		render(
+			<Component label="Test Toggle Group Control" help="Help text">
+				{ options }
+			</Component>
+		);
+
+		expect(
+			screen.getByRole( 'radiogroup', {
+				description: 'Help text',
+			} )
+		).toBeVisible();
+	} );
+
+	it( 'should accessibly associate the help text when isDeselectable', () => {
+		render(
+			<Component
+				label="Test Toggle Group Control"
+				help="Help text"
+				isDeselectable
+			>
+				{ options }
+			</Component>
+		);
+
+		expect(
+			screen.getByRole( 'group', {
+				description: 'Help text',
+			} )
+		).toBeVisible();
+	} );
+
 	describe( 'isDeselectable', () => {
 		describe( 'isDeselectable = false', () => {
 			it( 'should not be deselectable', async () => {