+++ packages/ui/src/collapsible-card/header.tsx
@@ -1,11 +1,12 @@
 import clsx from 'clsx';
-import { forwardRef } from '@wordpress/element';
+import { forwardRef, useMemo, useState } from '@wordpress/element';
 import { chevronDown } from '@wordpress/icons';
 import * as Card from '../card';
 import * as Collapsible from '../collapsible';
 import { Icon } from '../icon';
 import styles from './style.module.css';
 import focusStyles from '../utils/css/focus.module.css';
+import { HeaderDescriptionIdContext } from './context';
 import type { HeaderProps } from './types';
 
 /**
@@ -22,38 +23,54 @@ export const Header = forwardRef< HTMLDivElement, HeaderProps >(
 		{ children, className, render, ...restProps },
 		ref
 	) {
+		const [ descriptionId, setDescriptionId ] = useState< string >();
+
+		const contextValue = useMemo(
+			() => ( { setDescriptionId } ),
+			[ setDescriptionId ]
+		);
+
 		return (
-			<Collapsible.Trigger
-				className={ clsx( styles.header, className ) }
-				render={
-					<Card.Header
-						ref={ ref }
-						render={ render }
-						{ ...restProps }
-					/>
-				}
-				nativeButton={ false }
-			>
-				<div className={ styles[ 'header-content' ] }>{ children }</div>
-				<div
-					className={ clsx( styles[ 'header-trigger-positioner' ] ) }
+			<HeaderDescriptionIdContext.Provider value={ contextValue }>
+				<Collapsible.Trigger
+					className={ clsx( styles.header, className ) }
+					render={
+						<Card.Header
+							ref={ ref }
+							render={ render }
+							{ ...restProps }
+						/>
+					}
+					nativeButton={ false }
+					aria-describedby={ descriptionId }
 				>
+					<div className={ styles[ 'header-content' ] }>
+						{ children }
+					</div>
 					<div
 						className={ clsx(
-							styles[ 'header-trigger-wrapper' ],
-							// While the interactive trigger element is the whole header,
-							// the focus ring will be displayed only on the icon to visually
-							// emulate it being the button.
-							focusStyles[ 'outset-ring--focus-parent-visible' ]
+							styles[ 'header-trigger-positioner' ]
 						) }
 					>
-						<Icon
-							icon={ chevronDown }
-							className={ styles[ 'header-trigger' ] }
-						/>
+						<div
+							className={ clsx(
+								styles[ 'header-trigger-wrapper' ],
+								// While the interactive trigger element is the whole header,
+								// the focus ring will be displayed only on the icon to visually
+								// emulate it being the button.
+								focusStyles[
+									'outset-ring--focus-parent-visible'
+								]
+							) }
+						>
+							<Icon
+								icon={ chevronDown }
+								className={ styles[ 'header-trigger' ] }
+							/>
+						</div>
 					</div>
-				</div>
-			</Collapsible.Trigger>
+				</Collapsible.Trigger>
+			</HeaderDescriptionIdContext.Provider>
 		);
 	}
 );